+++ app/[locale]/projects/[projectId]/page.tsx
@@ -1,60 +1,85 @@
 'use client';
 
-import React from 'react';
-import {Breadcrumb, Breadcrumbs, Link} from "react-aria-components";
+import { useEffect, useRef, useState } from 'react';
+import { useParams } from 'next/navigation';
+import { useTranslations } from 'next-intl';
+import {
+  Breadcrumb,
+  Breadcrumbs,
+  Link
+} from "react-aria-components";
+import {
+  useProjectQuery
+} from '@/generated/graphql';
+
+// Components
 import PageHeader from "@/components/PageHeader";
-import {Card} from '@/components/Card/card';
-import {useTranslations} from "next-intl";
+import { Card } from '@/components/Card/card';
 import {
   ContentContainer,
-  LayoutWithPanel,
-  SidebarPanel
+  LayoutContainer
 } from "@/components/Container";
+import ErrorMessages from '@/components/ErrorMessages';
+
+interface FunderInterface {
+  name: string;
+  shortName: string;
+  id: number;
+  grantId: string;
+}
+
+interface ProjectMemberInterface {
+  fullname: string;
+  role: string[];
+  email: string;
+}
+
+interface ResearchOutputsInterface {
+  title: string;
+}
+
+interface ProjectOverviewInterface {
+  title: string;
+  startDate: string | null;
+  endDate: string | null;
+  funders: FunderInterface[];
+  projectMembers: ProjectMemberInterface[];
+  researchOutputs: ResearchOutputsInterface[];
+}
 
 const ProjectOverviewPage: React.FC = () => {
-  const t = useTranslations('ProjectOverview');
+  // Get projectId param
+  const params = useParams();
+  const { projectId } = params; // From route /projects/:projectId
+  const errorRef = useRef<HTMLDivElement | null>(null);
+  const [errors, setErrors] = useState<string[]>([]);
+  const [project, setProject] = useState<ProjectOverviewInterface>({
+    title: '',
+    startDate: null,
+    endDate: null,
+    funders: [],
+    projectMembers: [],
+    researchOutputs: []
+  });
 
+  // Localization keys
+  const ProjectOverview = useTranslations('ProjectOverview');
+  const Global = useTranslations('Global');
 
-  const project = {
-    title: "Coastal Ocean Processes of North Greenland",
-    start_date: "2023-07-18",
-    end_date: "2026-06-30",
-    funders: [
-      {
-        name: "Physical Oceanography",
-        shortname: "PO",
-        id: "GR-1810",
-        grantid: "PO-GR-1810"
-      }
-    ],
-    project_members: [
-      {
-        fullname: "Frederick Ice",
-        role: "PI",
-        email: "fred.ice@example.com"
-      },
-      {
-        fullname: "Jennifer Frost",
-        role: "PI",
-        email: "jfrost@example.com"
-      },
-      {
-        fullname: "Amelia Snow",
-        role: "Other",
-        email: "asnow@example.com"
-      }
-    ],
-    research_outputs: [
-      {
-        title: "North Greenland Coastal Process Data Set 1"
-      },
-      {
-        title: "Ocean Temperature Measurements 2023-2024"
-      },
-      {
-        title: "Coastal Mapping Analysis Results"
-      }
-    ],
+  // Get Project using projectId
+  const { data, loading, error } = useProjectQuery(
+    {
+      variables: { projectId: Number(projectId) },
+      notifyOnNetworkStatusChange: true
+    }
+  );
+
+  if (error) {
+    const errorMsg = ProjectOverview('messages.errorGettingProject');
+    setErrors(prev => [...prev, errorMsg]);
+  }
+
+  const project1 = {
     plans: [
       {
         id: "plan_123",
@@ -107,102 +132,139 @@ const ProjectOverviewPage: React.FC = () => {
     ]
   };
 
+  useEffect(() => {
+    // When data from backend changes, set project data in state
+    if (data && data.project) {
+      setProject({
+        title: data.project.title ?? '',
+        startDate: data.project?.startDate ? data.project.startDate : '',
+        endDate: data.project?.endDate ? data.project.endDate : '',
+        funders: data.project.funders
+          ?.filter((funder) => funder !== null) // Filter out null
+          .map((funder) => ({
+            id: Number(funder.id),
+            name: funder.affiliation?.displayName ?? '',
+            shortName: funder.affiliation?.name ?? '',
+            grantId: funder.grantId ?? '',
+          })) ?? [], // Provide a default empty array
+        projectMembers: data.project.contributors
+          ?.filter((member) => member !== null) // Filter out null
+          .map((member) => ({
+            fullname: `${member.givenName} ${member.surName}`,
+            email: member.email ?? '',
+            role: (member.contributorRoles ?? []).map((role) => role.label),
+          })) ?? [], // Provide a default empty array
+        researchOutputs: data.project.outputs
+          ?.filter((output) => output !== null) // Filter out null
+          .map((output) => ({
+            title: output.title ?? '',
+          })) ?? [], // Provide a default empty array
+      });
+    }
+  }, [data]);
+
+
+  if (loading) {
+    return <div>{Global('messaging.loading')}...</div>;
+  }
+
   return (
     <>
       <PageHeader
-        title={t('pageTitle')}
-        description={t('pageDescription')}
-        showBackButton={true}
+        title={ProjectOverview('pageTitle')}
+        description={ProjectOverview('pageDescription')}
+        showBackButton={false}
         breadcrumbs={
-          <Breadcrumbs aria-label={t('navigation')}>
-            <Breadcrumb><Link href="/">{t('home')}</Link></Breadcrumb>
+          <Breadcrumbs aria-label={ProjectOverview('navigation')}>
+            <Breadcrumb><Link href="/">{ProjectOverview('home')}</Link></Breadcrumb>
             <Breadcrumb><Link
-              href="/projects">{t('projects')}</Link></Breadcrumb>
+              href="/projects">{ProjectOverview('projects')}</Link></Breadcrumb>
           </Breadcrumbs>
         }
         actions={null}
         className="page-project-list"
       />
-      <LayoutWithPanel>
+      <ErrorMessages errors={errors} ref={errorRef} />
+      <LayoutContainer>
         <ContentContainer>
           <div className="project-overview">
             <section className="project-overview-item project-header"
-                     aria-labelledby="project-title">
-              <h2 id="project-title">{t('project')}</h2>
+              aria-labelledby="project-title">
+              <h2 id="project-title">{ProjectOverview('project')}</h2>
               <p className="project-overview-item-heading">
                 <strong>
                   {project.title}
                 </strong>
               </p>
               <p>
-                {t('dateRange', {
-                  startDate: (project.start_date),
-                  endDate: (project.end_date)
+                {ProjectOverview('dateRange', {
+                  startDate: (project.startDate),
+                  endDate: (project.endDate)
                 })}
               </p>
-              <Link href="/projects/proj_2425/project" aria-label={t('editProject')}>
-                {t('edit')}
+              <Link href={`/projects/${projectId}/project`} aria-label={ProjectOverview('editProject')}>
+                {ProjectOverview('edit')}
               </Link>
             </section>
 
 
             <section className="project-overview-item project-funders"
-                     aria-labelledby="funders-title">
-              <h2 id="funders-title">{t('funders')}</h2>
+              aria-labelledby="funders-title">
+              <h2 id="funders-title">{ProjectOverview('funders')}</h2>
               <p className="project-overview-item-heading">
                 <strong>
-                  {t('funderCount', {count: project.funders.length})}
+                  {ProjectOverview('funderCount', { count: project.funders.length })}
                 </strong>
               </p>
               <p>
                 {project.funders.map((funder, index) => (
                   <span key={funder.id} data-index={index}>
-                      {t('funderInfo', {
-                        name: funder.name,
-                        id: funder.grantid
-                      })}
-                    </span>
+                    {funder.grantId ? (ProjectOverview('funderInfo', {
+                      name: funder.name,
+                      id: funder.grantId
+                    })) : funder.name}
+                  </span>
                 ))}
               </p>
-              <Link href="/projects/proj_2425/funder" aria-label={t('editFunders')}>
-                {t('editFunderDetails')}
+              <Link href={`/projects/${projectId}/funder`} aria-label={ProjectOverview('editFunders')}>
+                {ProjectOverview('editFunderDetails')}
               </Link>
             </section>
 
             <section className="project-overview-item project-members"
-                     aria-labelledby="members-title">
-              <h2 id="members-title">{t('projectMembers')}</h2>
+              aria-labelledby="members-title">
+              <h2 id="members-title">{ProjectOverview('projectMembers')}</h2>
               <p className="project-overview-item-heading">
                 <strong>
-                  {t('memberCount', {count: project.project_members.length})}
+                  {ProjectOverview('memberCount', { count: project.projectMembers.length })}
                 </strong>
               </p>
               <p>
-                {project.project_members.map((member, index) => (
+                {project.projectMembers.map((member, index) => (
                   <span key={index}>
-                    {t('memberInfo', {
+                    {ProjectOverview('memberInfo', {
                       name: member.fullname,
-                      role: member.role
+                      role: member.role.join(', ')
                     })}
-                    {index < project.project_members.length - 1 ? '; ' : ''}
+                    {index < project.projectMembers.length - 1 ? '; ' : ''}
                   </span>
                 ))}
               </p>
-              <Link href="/projects/proj_2425/members" aria-label={t('editMembers')}>
-                {t('editProjectMembers')}
+              <Link href={`/projects/${projectId}/members`} aria-label={ProjectOverview('editMembers')}>
+                {ProjectOverview('editProjectMembers')}
               </Link>
             </section>
 
             <section className="project-overview-item research-outputs"
-                     aria-labelledby="outputs-title">
-              <h2 id="outputs-title">{t('researchOutputs')}</h2>
+              aria-labelledby="outputs-title">
+              <h2 id="outputs-title">{ProjectOverview('researchOutputs')}</h2>
               <p className="project-overview-item-heading">
                 <strong>
-                  {t('outputCount', {count: project.research_outputs.length})}
+                  {ProjectOverview('outputCount', { count: project.researchOutputs.length })}
                 </strong>
               </p>
-              <Link href="/projects/proj_2425/research-outputs" aria-label={t('editOutputs')}>
-                {t('editResearchOutputs')}
+              <Link href={`/projects/${projectId}/research-outputs`} aria-label={ProjectOverview('editOutputs')}>
+                {ProjectOverview('editOutputs')}
               </Link>
             </section>
 
@@ -211,40 +273,40 @@ const ProjectOverviewPage: React.FC = () => {
           <section className="plans" aria-labelledby="plans-title">
             <div className="plans-header plans-header-with-actions">
               <div className="">
-                <h2 id="plans-title">{t('plans')}</h2>
+                <h2 id="plans-title">{ProjectOverview('plans')}</h2>
               </div>
               <div className="actions" role="group"
-                   aria-label={t('planActions')}>
+                aria-label={ProjectOverview('planActions')}>
                 <Link
-                  href="/en-US/projects/proj_2425/dmp/upload"
+                  href={`/projects/${projectId}/dmp/upload`}
                   className="react-aria-Button react-aria-Button--secondary"
-                  aria-label={t('uploadPlan')}
+                  aria-label={ProjectOverview('uploadPlan')}
                 >
-                  {t('upload')}
+                  {ProjectOverview('upload')}
                 </Link>
                 <Link
-                  href="/en-US/projects/proj_2425/dmp/create"
+                  href={`/projects/${projectId}/dmp/create`}
                   className="react-aria-Button react-aria-Button--primary"
-                  aria-label={t('createNewPlan')}
+                  aria-label={ProjectOverview('createNewPlan')}
                 >
-                  {t('createNew')}
+                  {ProjectOverview('createNew')}
                 </Link>
               </div>
             </div>
-            {project.plans.map((plan) => (
+            {project1.plans.map((plan) => (
               <Card className="plan-item" key={plan.id}>
-                <p className="mb-1">{t('funder')}: {plan.funder_name}</p>
+                <p className="mb-1">{ProjectOverview('funder')}: {plan.funder_name}</p>
                 <h3 className="mt-0">{plan.template_name}</h3>
                 <div className="plan-sections mb-4">
                   <ul className="plan-sections-list"
-                      aria-label={t('planSections')}>
+                    aria-label={ProjectOverview('planSections')}>
                     {plan.sections.map((section) => (
                       <li key={section.id} className="plan-sections-list-item">
                         <Link href={section.link}>
                           {section.section_title}
                         </Link>
                         <span className="plan-sections-list-item-progress">
-                          {t('progress', {
+                          {ProjectOverview('progress', {
                             current: section.progress,
                             total: 3
                           })}
@@ -255,9 +317,9 @@ const ProjectOverviewPage: React.FC = () => {
                 </div>
                 <div className="plan-meta">
                   <p>
-                    {t('doi')}: {plan.doi} <br/>
-                    {t('lastUpdated')}: {plan.last_updated}<br/>
-                    {t('template')}: Arctic Data Center: NSF Polar Programs
+                    {ProjectOverview('doi')}: {plan.doi} <br />
+                    {ProjectOverview('lastUpdated')}: {plan.last_updated}<br />
+                    {ProjectOverview('template')}: Arctic Data Center: NSF Polar Programs
                   </p>
                 </div>
                 <div className="plan-footer">
@@ -265,27 +327,27 @@ const ProjectOverviewPage: React.FC = () => {
                     <Link
                       href="/plans/123"
                       className="plan-link download-link"
-                      aria-label={t('downloadPlan')}
+                      aria-label={ProjectOverview('downloadPlan')}
                     >
-                      <span className="link-text">{t('download')}</span>
+                      <span className="link-text">{ProjectOverview('download')}</span>
                       <span className="link-icon" aria-hidden="true">↓</span>
                     </Link>
                     <Link
                       href="/share"
                       className="plan-link share-link"
-                      aria-label={t('sharePlan')}
+                      aria-label={ProjectOverview('sharePlan')}
                     >
-                      <span className="link-text">{t('share')}</span>
+                      <span className="link-text">{ProjectOverview('share')}</span>
                       <span className="link-icon" aria-hidden="true">👤</span>
                     </Link>
                   </div>
                   <div className="plan-action">
                     <Link
-                      href="/projects/proj_2425/dmp/xxx"
+                      href={`/projects/${projectId}/dmp/xxx`}
                       className="react-aria-Button react-aria-Button--primary"
-                      aria-label={t('updatePlan')}
+                      aria-label={ProjectOverview('updatePlan')}
                     >
-                      {t('update')}
+                      {ProjectOverview('update')}
                     </Link>
                   </div>
                 </div>
@@ -293,8 +355,7 @@ const ProjectOverviewPage: React.FC = () => {
             ))}
           </section>
         </ContentContainer>
-        <SidebarPanel></SidebarPanel>
-      </LayoutWithPanel>
+      </LayoutContainer>
 
     </>
   );