+++ app/[locale]/template/page.tsx
@@ -27,6 +27,7 @@ import ErrorMessages from '@/components/ErrorMessages';
 // Hooks
 import { useScrollToTop } from '@/hooks/scrollToTop';
 
+import logECS from '@/utils/clientLogger';
 import { TemplateInterface, TemplateItemProps, } from '@/app/types';
 import styles from './orgTemplates.module.scss';
 
@@ -54,7 +55,7 @@ const TemplateListPage: React.FC = () => {
   const SelectTemplate = useTranslations('TemplateSelectTemplatePage');
 
   // Make graphql request for templates under the user's affiliation
-  const { data = {}, loading, error: queryError, refetch } = useTemplatesQuery({
+  const { data = {}, loading, error: queryError } = useTemplatesQuery({
     /* Force Apollo to notify React of changes. This was needed for when refetch is
     called and a re-render of data is necessary*/
     notifyOnNetworkStatusChange: true,
@@ -131,15 +132,18 @@ const TemplateListPage: React.FC = () => {
   useEffect(() => {
     if (queryError) {
       if (queryError instanceof ApolloError) {
-        // Trigger a refetch on error so page re-renders for apollo errors
-        refetch();
+        setErrors(prevErrors => [...prevErrors, queryError.message]);
+        logECS('error', 'queryError', {
+          error: queryError,
+          url: { path: '/template' }
+        });
       } else {
         // Safely access queryError.message
         setErrors(prev => [...prev, t('somethingWentWrong')]);
       }
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [queryError, refetch]);
+  }, [queryError]);
 
 
   useEffect(() => {